Always install with "--with-all-dependencies"
In the current case we have the follwoing failing because it installs v2
of a dependency and then a later bundle requires v1.

See https://github.com/doctrine/orm/discussions/10297

We can't really stay on top of these issues of external projects, so
allow updating/downgrading dependencies when installing bundles.